make changes in line with team  reviewer request
